refactor: update penalties schema and interfaces, removing unused interface file
