+++ packages/data/schema/system/src/model/system-settings/penalties/fees_config_schema.schema.ts
@@ -1,6 +1,16 @@
 //@ts-ignore
 import { Schema, model } from 'mongoose';
-import { IFeesConfig } from './interface';
+//import { IFeesConfig } from './interface';
+import { ICreatedByAdmin } from '@isyss-cdm/interface';
+
+export interface IFeesConfig extends Document {
+  id: string;
+  licenseType?: string;
+  baseFee?: number;
+  createdBy?: ICreatedByAdmin;
+  createdAt?: Date;
+  updatedAt?: Date;
+}
 
 const FeesConfigSchema = new Schema<IFeesConfig>(
   {
