+++ packages/data/schema/system/src/model/system-settings/penalties/penalty_config.schema.ts
@@ -1,6 +1,18 @@
 //@ts-ignore
 import { Schema, model } from 'mongoose';
-import { IPenaltyConfig } from './interface';
+// import { IPenaltyConfig } from './interface';
+import { ICreatedByAdmin } from '@isyss-cdm/interface';
+
+// Interfaces
+export interface IPenaltyConfig extends Document {
+  id: string;
+  licenseType?: string;
+  tiers?: Record<string, unknown>;
+  lockThreshold?: number;
+  createdBy?: ICreatedByAdmin;
+  createdAt?: Date;
+  updatedAt?: Date;
+}
 
 const PenaltyConfigSchema = new Schema<IPenaltyConfig>(
   {
